Move graph generator utilities, simplify code
